+++ src/Type/DateComparisonOperatorType.php
@@ -17,6 +17,7 @@ class DateComparisonOperatorType extends EnumType
     public const BEFORE = 'BEFORE';
     public const AFTER = 'AFTER';
     public const BETWEEN = 'BETWEEN';
+    public const NOT_BETWEEN = 'NOT_BETWEEN';
 
     public function __construct()
     {
@@ -26,6 +27,7 @@ class DateComparisonOperatorType extends EnumType
                 self::BEFORE => [],
                 self::AFTER => [],
                 self::BETWEEN => [],
+                self::NOT_BETWEEN => [],
             ],
         ];