+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v1.6.0 - (Unreleased)
  * Some minors bugs fixed
  * Support to use custom entity primary id
  * Subscriptions refactoring and some improvements
+ * Add support to use NOT_BETWEEN in date filters
 
 v1.5.0 - 2020-02-19
 ----